modular parser updated
